Move the 'getTimingsForPath' tests to store-based tests
