+++ src/test/unit/profile-data.test.js
@@ -16,8 +16,6 @@ import {
   getCallNodePathFromIndex,
   getSampleIndexClosestToTime,
   convertStackToCallNodePath,
-  invertCallstack,
-  getTimingsForPath,
   getSampleIndexToCallNodeIndex,
   getCallNodeIndexFromPath,
   getTreeOrderComparator,
@@ -36,14 +34,11 @@ import { ensureExists } from '../../utils/flow';
 import getCallNodeProfile from '../fixtures/profiles/call-nodes';
 import {
   getProfileFromTextSamples,
-  getMergedProfileFromTextSamples,
   getJsTracerTable,
 } from '../fixtures/profiles/processed-profile';
 import { funcHasRecursiveCall } from '../../profile-logic/transforms';
 
 import type { Thread, IndexIntoStackTable } from '../../types/profile';
-import type { Milliseconds } from '../../types/units';
-import type { BreakdownByCategory } from '../../profile-logic/profile-data';
 
 describe('unique-string-array', function() {
   const u = new UniqueStringArray(['foo', 'bar', 'baz']);
@@ -785,459 +780,6 @@ describe('convertStackToCallNodePath', function() {
     expect(callNodePath).toEqual([0, 1, 2, 3, 4]);
     callNodePath = convertStackToCallNodePath(thread, stack2);
     expect(callNodePath).toEqual([0, 1, 2, 3, 5]);
-  });
-});
-
-// Creates a BreakdownByCategory for the case where every category has just one
-// subcategory (the "Other" subcategory), so that it's easier to write the
-// reference structures.
-function withSingleSubcategory(
-  categoryBreakdown: Milliseconds[]
-): BreakdownByCategory {
-  return categoryBreakdown.map(value => ({
-    entireCategoryValue: value,
-    subcategoryBreakdown: [value],
-  }));
-}
-
-describe('getTimingsForPath in a non-inverted tree', function() {
-  function setup() {
-    const { profile, funcNamesDictPerThread } = getProfileFromTextSamples(`
-      A                  A             A             A              A
-      B                  B             B             B              B
-      Cjs                Cjs           Cjs           H[cat:Layout]  H[cat:Layout]
-      D                  D             F             I[cat:Idle]
-      Ejs[jit:baseline]  Ejs[jit:ion]  Ejs[jit:ion]
-    `);
-
-    const defaultCategory = profile.meta.categories.findIndex(
-      c => c.name === 'Other'
-    );
-    const thread = profile.threads[0];
-    const callNodeInfo = getCallNodeInfo(
-      thread.stackTable,
-      thread.frameTable,
-      thread.funcTable,
-      defaultCategory
-    );
-    const curriedGetTimingsForPath = path =>
-      getTimingsForPath(
-        path,
-        callNodeInfo,
-        profile.meta.interval,
-        false,
-        thread,
-        profile.meta.categories
-      );
-
-    return {
-      getTimingsForPath: curriedGetTimingsForPath,
-      funcNamesDict: funcNamesDictPerThread[0],
-    };
-  }
-
-  it('returns good timings for a root node', () => {
-    const {
-      getTimingsForPath,
-      funcNamesDict: { A },
-    } = setup();
-
-    // This is a root node: it should have no self time but all the total time.
-    const timings = getTimingsForPath([A]);
-    expect(timings).toEqual({
-      forPath: {
-        selfTime: {
-          value: 0,
-          breakdownByImplementation: null,
-          breakdownByCategory: null,
-        },
-        totalTime: {
-          value: 5,
-          breakdownByImplementation: { native: 2, baseline: 1, ion: 2 },
-          breakdownByCategory: withSingleSubcategory([1, 0, 1, 3, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript, ...]
-        },
-      },
-      forFunc: {
-        selfTime: {
-          value: 0,
-          breakdownByImplementation: null,
-          breakdownByCategory: null,
-        },
-        totalTime: {
-          value: 5,
-          breakdownByImplementation: { native: 2, baseline: 1, ion: 2 },
-          breakdownByCategory: withSingleSubcategory([1, 0, 1, 3, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript ...]
-        },
-      },
-      rootTime: 5,
-    });
-  });
-
-  it('returns good timings for a leaf node, also present in other stacks', () => {
-    const {
-      getTimingsForPath,
-      funcNamesDict: { A, B, Cjs, D, Ejs },
-    } = setup();
-
-    // This is a leaf node: it should have some self time and some total time
-    // holding the same value.
-    //
-    // This is also a JS node so it should have some js engine implementation
-    // implementations.
-    //
-    // The same func is also present in 2 different stacks so it should have
-    // different timings for the `forFunc` property.
-    const timings = getTimingsForPath([A, B, Cjs, D, Ejs]);
-    expect(timings).toEqual({
-      forPath: {
-        selfTime: {
-          value: 2,
-          breakdownByImplementation: { ion: 1, baseline: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 0, 2, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript, ...]
-        },
-        totalTime: {
-          value: 2,
-          breakdownByImplementation: { ion: 1, baseline: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 0, 2, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript, ...]
-        },
-      },
-      forFunc: {
-        selfTime: {
-          value: 3,
-          breakdownByImplementation: { ion: 2, baseline: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 0, 3, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript, ...]
-        },
-        totalTime: {
-          value: 3,
-          breakdownByImplementation: { ion: 2, baseline: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 0, 3, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript, ...]
-        },
-      },
-      rootTime: 5,
-    });
-  });
-
-  it('returns good timings for a node that has both children and self time', () => {
-    const {
-      getTimingsForPath,
-      funcNamesDict: { A, B, H },
-    } = setup();
-
-    // This is a node that has both children and some self time. So it should
-    // have some running time that's different than the self time.
-    const timings = getTimingsForPath([A, B, H]);
-    expect(timings).toEqual({
-      forPath: {
-        selfTime: {
-          value: 1,
-          breakdownByImplementation: { native: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 1, 0, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript ...]
-        },
-        totalTime: {
-          value: 2,
-          breakdownByImplementation: { native: 2 },
-
-          breakdownByCategory: withSingleSubcategory([1, 0, 1, 0, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript ...]
-        },
-      },
-      forFunc: {
-        selfTime: {
-          value: 1,
-          breakdownByImplementation: { native: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 1, 0, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript ...]
-        },
-        totalTime: {
-          value: 2,
-          breakdownByImplementation: { native: 2 },
-          breakdownByCategory: withSingleSubcategory([1, 0, 1, 0, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript ...]
-        },
-      },
-      rootTime: 5,
-    });
-  });
-});
-
-describe('getTimingsForPath for an inverted tree', function() {
-  function setup() {
-    const { profile, funcNamesDictPerThread } = getProfileFromTextSamples(`
-      A                  A             A             A              A
-      B                  B             B             B              B
-      Cjs                Cjs           Cjs           H[cat:Layout]  H[cat:Layout]
-      D                  D             F             I[cat:Idle]
-      Ejs[jit:baseline]  Ejs[jit:ion]  Ejs[jit:ion]
-    `);
-    const defaultCategory = profile.meta.categories.findIndex(
-      c => c.name === 'Other'
-    );
-    const thread = invertCallstack(profile.threads[0], defaultCategory);
-    // Now the profile should look like this:
-    //
-    // Ejs  Ejs  Ejs  I[cat:Idle]    H[cat:Layout]
-    // D    D    F    H[cat:Layout]  B
-    // Cjs  Cjs  Cjs  B              A
-    // B    B    B    A
-    // A    A    A
-
-    const callNodeInfo = getCallNodeInfo(
-      thread.stackTable,
-      thread.frameTable,
-      thread.funcTable,
-      defaultCategory
-    );
-    const curriedGetTimingsForPath = path =>
-      getTimingsForPath(
-        path,
-        callNodeInfo,
-        profile.meta.interval,
-        true,
-        thread,
-        profile.meta.categories
-      );
-
-    return {
-      getTimingsForPath: curriedGetTimingsForPath,
-      funcNamesDict: funcNamesDictPerThread[0],
-    };
-  }
-
-  it('returns good timings for a root node', () => {
-    const {
-      getTimingsForPath,
-      funcNamesDict: { Ejs },
-    } = setup();
-    const timings = getTimingsForPath([Ejs]);
-    expect(timings).toEqual({
-      forPath: {
-        selfTime: {
-          value: 3,
-          breakdownByImplementation: null,
-          breakdownByCategory: null,
-        },
-        totalTime: {
-          value: 3,
-          breakdownByImplementation: { ion: 2, baseline: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 0, 3, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript, ...]
-        },
-      },
-      forFunc: {
-        selfTime: {
-          value: 3,
-          breakdownByImplementation: { ion: 2, baseline: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 0, 3, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript, ...]
-        },
-        totalTime: {
-          value: 3,
-          breakdownByImplementation: { ion: 2, baseline: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 0, 3, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript, ...]
-        },
-      },
-      rootTime: 5,
-    });
-  });
-
-  it('returns good timings for a node present in several stacks without self time', () => {
-    const {
-      getTimingsForPath,
-      funcNamesDict: { Ejs, D, Cjs, B },
-    } = setup();
-    const timings = getTimingsForPath([Ejs, D, Cjs, B]);
-    expect(timings).toEqual({
-      forPath: {
-        selfTime: {
-          value: 0,
-          breakdownByImplementation: null,
-          breakdownByCategory: null,
-        },
-        totalTime: {
-          value: 2,
-          breakdownByImplementation: { ion: 1, baseline: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 0, 2, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript, ...]
-        },
-      },
-      forFunc: {
-        selfTime: {
-          value: 0,
-          breakdownByImplementation: null,
-          breakdownByCategory: null,
-        },
-        totalTime: {
-          value: 5,
-          breakdownByImplementation: {
-            ion: 2,
-            baseline: 1,
-            native: 2,
-          },
-          breakdownByCategory: withSingleSubcategory([1, 0, 1, 3, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript, ...]
-        },
-      },
-      rootTime: 5,
-    });
-  });
-
-  it('returns good timings for a node present in several stacks with self time', () => {
-    const {
-      getTimingsForPath,
-      funcNamesDict: { I, H },
-    } = setup();
-
-    // Select the function as a root node
-    let timings = getTimingsForPath([H]);
-    expect(timings).toEqual({
-      forPath: {
-        selfTime: {
-          value: 1,
-          breakdownByImplementation: null,
-          breakdownByCategory: null,
-        },
-        totalTime: {
-          value: 1,
-          breakdownByImplementation: { native: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 1, 0, 0, 0, 0, 0]), // [Idle, Other, Layout]
-        },
-      },
-      forFunc: {
-        selfTime: {
-          value: 1,
-          breakdownByImplementation: { native: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 1, 0, 0, 0, 0, 0]), // [Idle, Other, Layout]
-        },
-        totalTime: {
-          value: 2,
-          breakdownByImplementation: { native: 2 },
-          breakdownByCategory: withSingleSubcategory([1, 0, 1, 0, 0, 0, 0, 0]), // [Idle, Other, Layout]
-        },
-      },
-      rootTime: 5,
-    });
-
-    // Select the same function, but this time when it's not a root node
-    timings = getTimingsForPath([I, H]);
-    expect(timings).toEqual({
-      forPath: {
-        selfTime: {
-          value: 0,
-          breakdownByImplementation: null,
-          breakdownByCategory: null,
-        },
-        totalTime: {
-          value: 1,
-          breakdownByImplementation: { native: 1 },
-          breakdownByCategory: withSingleSubcategory([1, 0, 0, 0, 0, 0, 0, 0]), // [Idle]
-        },
-      },
-      forFunc: {
-        selfTime: {
-          value: 1,
-          breakdownByImplementation: { native: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 1, 0, 0, 0, 0, 0]), // [Idle, Other, Layout]
-        },
-        totalTime: {
-          value: 2,
-          breakdownByImplementation: { native: 2 },
-          breakdownByCategory: withSingleSubcategory([1, 0, 1, 0, 0, 0, 0, 0]), // [Idle, Other, Layout]
-        },
-      },
-      rootTime: 5,
-    });
-  });
-
-  it('returns good timings for a leaf node', () => {
-    const {
-      getTimingsForPath,
-      funcNamesDict: { H, B, A },
-    } = setup();
-    const timings = getTimingsForPath([H, B, A]);
-    expect(timings).toEqual({
-      forPath: {
-        selfTime: {
-          value: 0,
-          breakdownByImplementation: null,
-          breakdownByCategory: null,
-        },
-        totalTime: {
-          value: 1,
-          breakdownByImplementation: { native: 1 },
-          breakdownByCategory: withSingleSubcategory([0, 0, 1, 0, 0, 0, 0, 0]), // [Idle, Other, Layout]
-        },
-      },
-      forFunc: {
-        selfTime: {
-          value: 0,
-          breakdownByImplementation: null,
-          breakdownByCategory: null,
-        },
-        totalTime: {
-          value: 5,
-          breakdownByImplementation: { native: 2, ion: 2, baseline: 1 },
-          breakdownByCategory: withSingleSubcategory([1, 0, 1, 3, 0, 0, 0, 0]), // [Idle, Other, Layout, JavaScript]
-        },
-      },
-      rootTime: 5,
-    });
-  });
-});
-
-describe('getTimingsForPath for a diffing track', function() {
-  function setup() {
-    const { profile, funcNamesDictPerThread } = getMergedProfileFromTextSamples(
-      `
-      A              A  A
-      B              B  C
-      D[cat:Layout]  E  F
-    `,
-      `
-      A                  A  A
-      B                  B  B
-      G[cat:JavaScript]  I  E
-    `
-    );
-    const defaultCategory = profile.meta.categories.findIndex(
-      c => c.name === 'Other'
-    );
-    const thread = profile.threads[2];
-
-    const callNodeInfo = getCallNodeInfo(
-      thread.stackTable,
-      thread.frameTable,
-      thread.funcTable,
-      defaultCategory
-    );
-    const curriedGetTimingsForPath = path =>
-      getTimingsForPath(
-        path,
-        callNodeInfo,
-        profile.meta.interval,
-        false /* inverted tree */,
-        thread,
-        profile.meta.categories
-      );
-
-    return {
-      getTimingsForPath: curriedGetTimingsForPath,
-      funcNamesDictPerThread,
-    };
-  }
-
-  it('computes the right breakdowns', () => {
-    const {
-      getTimingsForPath,
-      funcNamesDictPerThread: [{ A }],
-    } = setup();
-    const timings = getTimingsForPath([A]);
-    expect(timings.forPath).toEqual({
-      selfTime: {
-        breakdownByCategory: null,
-        breakdownByImplementation: null,
-        value: 0,
-      },
-      totalTime: {
-        breakdownByCategory: withSingleSubcategory([0, 0, -1, 1, 0, 0, 0, 0]), // Idle, Other, Layout, JavaScript, etc.
-        breakdownByImplementation: {
-          native: 0,
-        },
-        value: 0,
-      },
-    });
   });
 });
